function __e_acsl_offset now returns size_t (fixed gitlab issue #10)

Closes #10

See merge request frama-c/e-acsl!194

+++ src/plugins/e-acsl/share/e-acsl/bittree_model/e_acsl_bittree_mmodel.c
@@ -270,7 +270,7 @@ void* base_addr(void* ptr) {
 }
 
 /* return the offset of `ptr` within its block */
-int offset(void* ptr) {
+size_t offset(void* ptr) {
   bt_block * tmp = bt_find(ptr);
   vassert(tmp != NULL, "\\offset of unallocated memory", NULL);
   return ((uintptr_t)ptr - tmp->ptr);
